move classes to callback-common module
