+++ library/extensions/kmp-tor-ext-callback-controller-common/src/commonMain/kotlin/io/matthewnelson/kmp/tor/ext/callback/controller/common/ResultExt.kt
@@ -16,6 +16,7 @@
 package io.matthewnelson.kmp.tor.ext.callback.controller.common
 
 import io.matthewnelson.kmp.tor.common.annotation.InternalTorApi
+import io.matthewnelson.kmp.tor.ext.callback.common.TorCallback
 
 @InternalTorApi
 @Suppress("nothing_to_inline")
