+++ library/extensions/kmp-tor-ext-callback-controller-common/src/commonMain/kotlin/io/matthewnelson/kmp/tor/ext/callback/controller/common/control/usecase/CallbackTorControlOwnershipDrop.kt
@@ -15,8 +15,8 @@
  **/
 package io.matthewnelson.kmp.tor.ext.callback.controller.common.control.usecase
 
-import io.matthewnelson.kmp.tor.ext.callback.controller.common.Task
-import io.matthewnelson.kmp.tor.ext.callback.controller.common.TorCallback
+import io.matthewnelson.kmp.tor.ext.callback.common.Task
+import io.matthewnelson.kmp.tor.ext.callback.common.TorCallback
 
 /**
  * "DROPOWNERSHIP" CRLF
